@@ -12,20 +12,22 @@
   ],
   "license": "MIT",
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "ext-json": "*",
     "ext-pdo": "*",
-    "setbased/exception": "^2.1.2",
-    "setbased/helper-cast": "^2.1.1",
-    "setbased/helper-code-store-php": "^2.4.2",
-    "setbased/php-stratum-backend": "^5.3.0",
-    "setbased/php-stratum-common": "^6.2.0",
-    "setbased/php-stratum-middle": "^5.9.0"
+    "setbased/exception": "^2.3.0",
+    "setbased/helper-cast": "^3.0.0",
+    "setbased/helper-code-store-php": "^2.6.0",
+    "setbased/php-stratum-backend": "^6.0.0",
+    "setbased/php-stratum-common": "^6.3.1",
+    "setbased/php-stratum-middle": "^5.11.0"
   },
+  "minimum-stability": "dev",
+  "prefer-stable": true,
   "require-dev": {
-    "phpunit/phpunit": "^9.5.0",
-    "phing/phing": "^2.16.3",
-    "setbased/php-stratum": "^6.1.0"
+    "phing/phing": "^3.0.0-RC4",
+    "phpunit/phpunit": "^9.5.28",
+    "setbased/php-stratum": "^6.3.1"
   },
   "autoload": {
     "psr-4": {
@@ -39,6 +41,9 @@
   },
   "config": {
     "bin-dir": "bin/",
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "phing/phing-composer-configurator": true
+    }
   }
 }
