@@ -44,10 +44,10 @@ class SqlitePdoWorker
   /**
    * Writes a file in two phase to the filesystem.
    *
-   * First write the data to a temporary file (in the same directory) and than renames the temporary file. If the file
+   * First write the data to a temporary file (in the same directory) and then renames the temporary file. If the file
    * already exists and its content is equal to the data that must be written no action  is taken. This has the
    * following advantages:
-   * * In case of some write error (e.g. disk full) the original file is kept in tact and no file with partially data
+   * * In case of some write error (e.g. disk full) the original file is kept intact and no file with partially data
    * is written.
    * * Renaming a file is atomic. So, running processes will never read a partially written data.
    *
@@ -60,7 +60,7 @@ class SqlitePdoWorker
     if (file_exists($filename))
     {
       $old_data = file_get_contents($filename);
-      if ($data==$old_data) $write_flag = false;
+      if ($data===$old_data) $write_flag = false;
     }
 
     if ($write_flag)
