@@ -13,7 +13,10 @@ use Symfony\Component\Console\Tester\ApplicationTester;
 class StratumApplicationTest extends TestCase
 {
   //--------------------------------------------------------------------------------------------------------------------
-  public function testExecute()
+  /**
+   * Test the stratum command.
+   */
+  public function testExecute(): void
   {
     $application = new Stratum();
     $application->setAutoExit(false);
