@@ -12,7 +12,7 @@ class RowsTest extends DataLayerTestCase
   /**
    * Stored routine with designation type rows must return an empty array when no rows are selected.
    */
-  public function testRows0()
+  public function testRows0(): void
   {
     $ret = $this->dataLayer->tstTestRows1(0);
     self::assertIsArray($ret);
@@ -23,7 +23,7 @@ class RowsTest extends DataLayerTestCase
   /**
    * Stored routine with designation type rows must return an array with 1 row when only 1 row is selected.
    */
-  public function testRows1()
+  public function testRows1(): void
   {
     $ret = $this->dataLayer->tstTestRows1(1);
     self::assertIsArray($ret);
@@ -34,7 +34,7 @@ class RowsTest extends DataLayerTestCase
   /**
    * Stored routine with designation type rows must return an array with 3 rows when 3 rows are selected.
    */
-  public function testRows3()
+  public function testRows3(): void
   {
     $ret = $this->dataLayer->tstTestRows1(3);
     self::assertIsArray($ret);
