@@ -14,7 +14,7 @@ class Singleton0Test extends DataLayerTestCase
   /**
    * Stored routine with designation type singleton0 must return null.
    */
-  public function test01()
+  public function test01(): void
   {
     $value = $this->dataLayer->tstTestSingleton0a(0);
     self::assertNull($value);
@@ -24,7 +24,7 @@ class Singleton0Test extends DataLayerTestCase
   /**
    * Stored routine with designation type singleton0 must return 1 value.
    */
-  public function test02()
+  public function test02(): void
   {
     $value = $this->dataLayer->tstTestSingleton0a(1);
     self::assertIsInt($value);
@@ -34,7 +34,7 @@ class Singleton0Test extends DataLayerTestCase
   /**
    * An exception must be thrown when a stored routine with designation type singleton0 returns more than 1 rows.
    */
-  public function test03()
+  public function test03(): void
   {
     $this->expectException(ResultException::class);
     $this->dataLayer->tstTestSingleton0a(2);
@@ -44,7 +44,7 @@ class Singleton0Test extends DataLayerTestCase
   /**
    * Stored routine with designation type singleton0 and return type bool must return false when selecting 0 rows.
    */
-  public function test11()
+  public function test11(): void
   {
     $value = $this->dataLayer->tstTestSingleton0b(0, 1);
     self::assertFalse($value);
@@ -55,7 +55,7 @@ class Singleton0Test extends DataLayerTestCase
    * Stored routine with designation type singleton0 and return type bool|null must return false when selecting 1 row
    * with null value.
    */
-  public function test12()
+  public function test12(): void
   {
     $value = $this->dataLayer->tstTestSingleton0b(1, null);
     self::assertFalse($value);
@@ -66,7 +66,7 @@ class Singleton0Test extends DataLayerTestCase
    * Stored routine with designation type singleton0 and return type bool must return false when selecting 1 row
    * with empty value.
    */
-  public function test13()
+  public function test13(): void
   {
     $value = $this->dataLayer->tstTestSingleton0b(1, 0);
     self::assertFalse($value);
@@ -77,7 +77,7 @@ class Singleton0Test extends DataLayerTestCase
    * Stored routine with designation type singleton0 and return type bool must return true when selecting 1 row
    * with non-empty value.
    */
-  public function test14()
+  public function test14(): void
   {
     $value = $this->dataLayer->tstTestSingleton0b(1, 123);
     self::assertTrue($value);
@@ -86,9 +86,9 @@ class Singleton0Test extends DataLayerTestCase
   //--------------------------------------------------------------------------------------------------------------------
   /**
    * An exception must be thrown when a stored routine with designation type singleton0  and return type bool returns
-   * more than 1 rows.
+   * more than 1 row.
    */
-  public function test15()
+  public function test15(): void
   {
     $this->expectException(ResultException::class);
     $this->dataLayer->tstTestSingleton0b(2, 1);
