@@ -14,7 +14,7 @@ class Singleton1Test extends DataLayerTestCase
   /**
    * Stored routine with designation type singleton1 must return 1 value and 1 value only.
    */
-  public function test01()
+  public function test01(): void
   {
     $ret = $this->dataLayer->tstTestSingleton1a(1);
     self::assertEquals(1, $ret);
@@ -24,7 +24,7 @@ class Singleton1Test extends DataLayerTestCase
   /**
    * An exception must be thrown when a stored routine with designation type singleton1 returns 0 rows.
    */
-  public function test02()
+  public function test02(): void
   {
     $this->expectException(ResultException::class);
     $this->dataLayer->tstTestSingleton1a(0);
@@ -34,7 +34,7 @@ class Singleton1Test extends DataLayerTestCase
   /**
    * An exception must be thrown when a stored routine with designation type singleton1 returns more than 1 rows.
    */
-  public function test03()
+  public function test03(): void
   {
     $this->expectException(ResultException::class);
     $this->dataLayer->tstTestSingleton1a(2);
@@ -44,7 +44,7 @@ class Singleton1Test extends DataLayerTestCase
   /**
    *  An exception must be thrown when a stored routine with designation type singleton1 return type bool returns 0 rows.
    */
-  public function test11()
+  public function test11(): void
   {
     $this->expectException(ResultException::class);
     $value = $this->dataLayer->tstTestSingleton1b(0, 1);
@@ -56,7 +56,7 @@ class Singleton1Test extends DataLayerTestCase
    * Stored routine with designation type singleton1 and return type bool must return false when selecting 1 row
    * with null value.
    */
-  public function test12()
+  public function test12(): void
   {
     $value = $this->dataLayer->tstTestSingleton1b(1, null);
     self::assertFalse($value);
@@ -67,7 +67,7 @@ class Singleton1Test extends DataLayerTestCase
    * Stored routine with designation type singleton1 and return type bool must return false when selecting 1 row
    * with empty value.
    */
-  public function test13()
+  public function test13(): void
   {
     $value = $this->dataLayer->tstTestSingleton1b(1, 0);
     self::assertFalse($value);
@@ -78,7 +78,7 @@ class Singleton1Test extends DataLayerTestCase
    * Stored routine with designation type singleton1 and return type bool must return true when selecting 1 row
    * with non-empty value.
    */
-  public function test14()
+  public function test14(): void
   {
     $value = $this->dataLayer->tstTestSingleton1b(1, 123);
     self::assertTrue($value);
@@ -87,9 +87,9 @@ class Singleton1Test extends DataLayerTestCase
   //--------------------------------------------------------------------------------------------------------------------
   /**
    * An exception must be thrown when a stored routine with designation type singleton1 and return type bool returns
-   * more than 1 rows.
+   * more than 1 row.
    */
-  public function test15()
+  public function test15(): void
   {
     $this->expectException(ResultException::class);
     $this->dataLayer->tstTestSingleton1b(2, 1);
